refactor(model): controller에서 model 분리
model 메서드에서 view 메서드를 콜백 함수로 받아 실행하도록 함.
